chore(deps): bump apis-acdhch-default-settings to v1.1.1
apis-core v0.26.0 changes how menu items are added to the relations
menu: the items are not hardcoded in the `base.html` template, but the
apps are responsible for overriding the `base.html` template and adding
them. therefore the order the apps are listed in INSTALLED_APPS is
important. apis-acdhch-default-settings v1.1.1 changes this order.
See commit 5f27f04146228d255a9e46dd53bb36e9947542a2 in the apis-core
repository.